requirements.txt
anyio==4.7.0
arrow==1.3.0
cachetools==5.5.2
certifi==2024.12.14
charset-normalizer==3.4.0
gspread==6.2.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
httplib2==0.22.0
httpx==0.24.1
oauth2client==4.1.3
openai==1.88.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.2.1
python-telegram-bot==20.3
requests==2.32.3
tqdm==4.67.1
